Fix  #2149 - Error 403 on admin-ajax requests and WooCommerce nonces (PR #2208)
In some cases, on WooCommerce websites with plugins, e.g. WooCommerce Mailchimp Discount, that use AJAX requests (admin-ajax.php), those fail with an error 403.

This is related to nonces and the way, when WooCommerce is installed, they are produced.

The issue has to do with the fact that WooCommerce uses the following filter, which is defined in the wp_create_nonce() function:

$uid = apply_filters( 'nonce_user_logged_out', $uid, $action );

When a user is logged out, WooCommerce sets the user ID for the non-logged in user to a random one generated with Password Hash.

We identified 2 nonces till now:

"Load more" AJAX functionality of the Newspaper theme is affected.
WooCommerce MailChimp Discount.